console: Move recursiveMap utility to lib
